refactor: organise feature and ui directories
