+++ src/features/nft-view/actions/Action/Action.module.scss
@@ -1,5 +1,5 @@
-@use '../../../../../zUI/src/styles/typography';
-@use '../../../../../zUI/src/styles/colors';
+@use '../../../../../../zUI/src/styles/typography';
+@use '../../../../../../zUI/src/styles/colors';
 
 .Container {
 	> span {
