+++ src/features/nft-view/actions/Actions/Actions.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../../zUI/src/styles/colors';
+@use '../../../../../../zUI/src/styles/colors';
 
 .Container {
 	display: flex;
