+++ src/features/nft-view/history/HistoryItem/HistoryItem.tsx
@@ -1,5 +1,5 @@
 //- Types Imports
-import { DomainEvents } from '../../../lib/types/events';
+import { DomainEvents } from '../../../../lib/types/events';
 
 //- Library Imports
 import moment from 'moment';
@@ -14,8 +14,8 @@ import {
 } from '@zero-tech/zns-sdk';
 
 //- Utils Imports
-import { truncateAddress } from '../../../lib/util/domains/domains';
-import { formatEthers } from '../../../lib/util/number/number';
+import { truncateAddress } from '../../../../lib/util/domains/domains';
+import { formatEthers } from '../../../../lib/util/number/number';
 
 type HistoryItemProps = {
 	item: DomainEvents;
