+++ src/features/subdomain-view/subdomain-table/SubdomainTable.constants.ts
