+++ src/features/subdomain-view/subdomain-table/SubdomainTableCard/SubdomainTableCard.module.scss
@@ -1,6 +1,6 @@
-@use '../../../../../zUI/src/styles/dimensions';
-@use '../../../../../zUI/src/styles/typography';
-@use '../../../../../zUI/src/styles/colors';
+@use '../../../../../../zUI/src/styles/dimensions';
+@use '../../../../../../zUI/src/styles/typography';
+@use '../../../../../../zUI/src/styles/colors';
 
 .Container {
 	display: flex;
