+++ src/features/subdomain-view/subdomain-table/SubdomainTableCard/SubdomainTableCard.tsx
@@ -6,17 +6,17 @@ import { ethers } from 'ethers';
 import { TokenPriceInfo } from '@zero-tech/zns-sdk';
 
 //- Hook Imports
-import { useDomainMetadata } from '../../../lib/hooks/useDomainMetadata';
-import { useDomainMetrics } from '../../../lib/hooks/useDomainMetrics';
-import { formatEthers, formatNumber } from '../../../lib/util/number/number';
-import { useBuyNowPrice } from '../../../lib/hooks/useBuyNowPrice';
+import { useDomainMetadata } from '../../../../lib/hooks/useDomainMetadata';
+import { useDomainMetrics } from '../../../../lib/hooks/useDomainMetrics';
+import { formatEthers, formatNumber } from '../../../../lib/util/number/number';
+import { useBuyNowPrice } from '../../../../lib/hooks/useBuyNowPrice';
 
 //- Components Imports
 import Button from '@zero-tech/zui/src/components/Button/index';
-import TableCard from '../../ui/TableCard/TableCard';
+import TableCard from '../../../ui/TableCard/TableCard';
 
 //- Constants Imports
-import { ModalType } from '../../../lib/constants/modals';
+import { ModalType } from '../../../../lib/constants/modals';
 
 //-Styles Imports
 import styles from './SubdomainTableCard.module.scss';
